requirements.txt
Mako==1.0.2
MarkupSafe==0.23
Pillow==3.0.0
appdirs==1.4.0
chainer==1.3.2
chainer-cuda-deps==1.1.0.1
decorator==4.0.4
filelock==2.0.4
nose==1.3.7
numpy==1.10.0.post2
protobuf==2.6.1
py==1.4.30
pycuda==2015.1.3
pytest==2.8.2
pytools==2015.1.3
scikit-cuda==0.5.0
six==1.10.0
